Extract reaction formatting into shared utility
Move formatReactionsShort and formatReactionsDetailed from chat.tsx and
search-messages.tsx into a shared reactions.ts module to eliminate duplication.
